Pull out table_validator
